index.html
CSS (Cascading Style Sheets) is a vital tool for web developers to enhance the appearance and visual aesthetics of their web pages. CSS allows developers to control the layout, styling, and formatting of HTML elements. There are three primary types of CSS: inline, internal, and external CSS which we are going to study in this file.