Create index.html
The display property is used to specify how an element is shown on a web page. Every HTML element has a default display value, depending on what type of element it is. The default display value for most elements is block or inline. The display property changes the default display behavior of HTML elements.