Create index.html
The box-shadow property enables you to cast a drop shadow from the frame of almost any element, while the outline draws a border-like shape around the boundaries of an element. This means the outline doesn't take up additional space in your design, won't affect the size of an element, and doesn't cause layout shifts when it appears.