Create index.html
The object-fit and object-position properties in CSS are used to control the content of replaced elements, such as images or videos. Object-fit. Specifies how an <img> or <video> should be resized to fit its container. object-position. Specifies how an <img> or <video> should be positioned with x/y coordinates inside its "own content box"