Create script.js
JavaScript is a programming language that adds interactivity to your website. In this. We will learn the nomenclature of the javascript variable, the creation of the variable, how to print the variable, how to store numbers characters, and strings in a variable, how to get the type of that variable, how to add or do operations on the variable also types of datatype. in javascript that is primitive and object datatype.