Create index.js
A JavaScript function runs when it is “called” by some part of your code. Syntax: The basic syntax to create a function in JavaScript is shown below. To create a function in JavaScript, we have first to use the keyword function, separated by the name of the function and parameters within the parenthesis