Create index.html
In CSS, a class is used to group more than one element, whereas an ID is used to identify a single element. A class selector is therefore used to style multiple HTML elements of the same class, while an ID selector is used to style one HTML element. You can recognize a class selector by the period (“.”). To select a class in JavaScript, you can use the document. querySelector() method along with the class selector. const element = document. querySelector('.')