Create index.html
1.Adding elements to arrays
Use the push() method to add an item to the end of an array, or the unshift() method to add an item to the beginning of an array. 
2.Removing elements from arrays
Use the pop() method to remove the last item from an array, or the shift() method to remove the first item from an array. Both methods modify the original array. You can also use the splice() method to remove an element from an array at a specific index. 
3.Adding elements to HTML
Use the createElement() method to create an HTML element, or the appendChild() and append() methods to add an element as the last child of a parent element. You can also use the insertBefore() method to insert an element before another element. 
4.Removing elements from HTML
Use the remove() or removeChild() methods to remove an element from the Document Object Model (DOM).